# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/spotify/currently-playing": {
      "get": {
        "summary": "Currently Playing",
        "description": "Get the currently playing track from user's spotify",
        "operationId": "currently_playing_spotify_currently_playing_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Track"
                }
              }
            }
          },
          "204": {
            "description": "Nothing is currently playing on user's spotify"
          },
          "502": {
            "description": "The Spotify API returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyErrorMessage"
                },
                "example": {
                  "error": "SpotifyError",
                  "message": "Authorization with access token unsuccessful"
                }
              }
            }
          }
        }
      }
    },
    "/spotify/last-played": {
      "get": {
        "summary": "Last Played",
        "description": "Get the last played track from user's spotify",
        "operationId": "last_played_spotify_last_played_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Track"
                }
              }
            }
          },
          "204": {
            "description": "No recent spotify track available"
          },
          "502": {
            "description": "The Spotify API returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpotifyErrorMessage"
                },
                "example": {
                  "error": "SpotifyError",
                  "message": "Authorization with access token unsuccessful"
                }
              }
            }
          }
        }
      }
    },
    "/spotify/top/{type}": {
      "get": {
        "summary": "Top Type",
        "description": "Get the top `type` from user's spotify",
        "operationId": "top_type_spotify_top__type__get",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "artists",
                "tracks"
              ],
              "type": "string",
              "title": "Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Track"
                      }
                    },
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TopArtist"
                      }
                    }
                  ],
                  "title": "Response Top Type Spotify Top  Type  Get"
                }
              }
            }
          },
          "204": {
            "description": "No top spotify data available"
          },
          "502": {
            "description": "The Spotify API returned an error",
            "content": {
              "application/json": {
                "example": {
                  "error": "SpotifyError",
                  "message": "Authorization with access token unsuccessful"
                },
                "schema": {
                  "$ref": "#/components/schemas/SpotifyErrorMessage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/books/currently-reading": {
      "get": {
        "summary": "Currently Reading",
        "operationId": "currently_reading_books_currently_reading_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HardcoverBook"
                }
              }
            }
          },
          "204": {
            "description": "No active currently-reading book"
          },
          "502": {
            "description": "The Hardcover API returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HardcoverErrorMessage"
                },
                "example": {
                  "error": "HardcoverError",
                  "message": "The upstream hardcover service returned an invalid response"
                }
              }
            }
          }
        }
      }
    },
    "/projects/featured": {
      "get": {
        "summary": "Featured Projects",
        "operationId": "featured_projects_projects_featured_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "default": 2,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeaturedProject"
                  },
                  "title": "Response Featured Projects Projects Featured Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/healthcheck": {
      "get": {
        "summary": "Healthcheck",
        "operationId": "healthcheck_healthcheck_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthcheckResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FeaturedProject": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "href": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Href"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/ProjectLink"
            },
            "type": "array",
            "title": "Links"
          },
          "accent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "href"
        ],
        "title": "FeaturedProject"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HardcoverBook": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "image_dominant_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Dominant Color"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "link": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Link"
          }
        },
        "type": "object",
        "required": [
          "title",
          "author",
          "link"
        ],
        "title": "HardcoverBook"
      },
      "HardcoverErrorMessage": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "HardcoverErrorMessage"
      },
      "HealthcheckResponse": {
        "properties": {
          "condition": {
            "type": "string",
            "title": "Condition"
          }
        },
        "type": "object",
        "required": [
          "condition"
        ],
        "title": "HealthcheckResponse"
      },
      "ProjectLink": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "href": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Href"
          },
          "external": {
            "type": "boolean",
            "title": "External",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "label",
          "href"
        ],
        "title": "ProjectLink"
      },
      "SpotifyErrorMessage": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "SpotifyErrorMessage"
      },
      "TopArtist": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          },
          "image": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "name",
          "url",
          "image"
        ],
        "title": "TopArtist",
        "description": "Dataclass representing an artist returned from the spotify's top artists response."
      },
      "Track": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "artists": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Artists"
          },
          "track_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Track Url"
          },
          "is_playing": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Playing"
          },
          "album_name": {
            "type": "string",
            "title": "Album Name"
          },
          "album_image": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Album Image"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "progress_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Ms"
          }
        },
        "type": "object",
        "required": [
          "name",
          "artists",
          "track_url",
          "album_name",
          "album_image"
        ],
        "title": "Track",
        "description": "Dataclass representing the a track returned from the spotify api.\n\nMost of the data from the original response isn't included here since\nwe're only interested in data our application actually requires."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}